requirements.txt
sentence-transformers==2.2.2
numpy==1.25.0
fastapi==0.98.0
uvicorn==0.22.0
gunicorn==20.1.0
hypercorn==0.14.3
fastapi-cache2==0.2.1